refactor(iso)!: remove loader.useStream; streaming consumed via loader.View
The bare useStream hook orphaned its SSR markup during hydration; .View routes
through LoaderHost's Suspense + useId machinery instead. Removes
use-loader-stream + its hook/types (StreamStatus re-homed to use-loader-runner),
reworks the live guard (live loaders require loader.View(render, { initial,
reduce }); .Boundary no longer throws), and ports the accumulation/status tests
onto the .View path. iso suite green (555).
